refactor!: remove the deprecated `cargo-packager-config` crate
