+++ .changes/config-removed.md
@@ -1,0 +1,5 @@
+---
+"cargo-packager": "minor"
+---
+
+Remove the deprecated `cargo-packager-config` dependency.
