--- crates/config/CHANGELOG.md
@@ -1,5 +1,0 @@
-# Changelog
-
-## \[0.2.0]
-
-- [`1809f10`](https://www.github.com/crabnebula-dev/cargo-packager/commit/1809f10b4fd1720fd740196f67c3c980ade0a6bd) Add `Config.enabled` field control whether this config is enabled or not
