--- crates/config/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "cargo-packager-config"
-version = "0.2.0"
-description = "Config types for cargo-packager."
-authors = [
-  "CrabNebula Ltd.",
-  "Tauri Programme within The Commons Conservancy",
-  "George Burton <burtonageo@gmail.com>"
-]
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-
-[features]
-clap = [ "dep:clap" ]
-
-[dependencies]
-serde = { workspace = true }
-schemars = { workspace = true }
-clap = { workspace = true, optional = true }
-strsim = "0.10"
