--- crates/config/README.md
@@ -1,7 +1,0 @@
-# cargo-packager-config
-
-Config types for [`cargo-packager`](../packager/).
-
-## Licenses
-
-MIT or MIT/Apache 2.0 where applicable.
