+++ crates/config_schema_generator/build.rs
@@ -10,12 +10,10 @@ use std::{
 };
 
 pub fn main() -> Result<(), Box<dyn Error>> {
-    let schema = schemars::schema_for!(cargo_packager_config::Config);
+    let schema = schemars::schema_for!(cargo_packager::Config);
     let schema_str = serde_json::to_string_pretty(&schema).unwrap();
     let crate_dir = PathBuf::from(std::env::var("CARGO_MANIFEST_DIR")?);
-    for file in ["../packager/schema.json", "../config/schema.json"] {
-        let mut schema_file = BufWriter::new(File::create(crate_dir.join(file))?);
-        write!(schema_file, "{schema_str}")?;
-    }
+    let mut schema_file = BufWriter::new(File::create(crate_dir.join("../packager/schema.json"))?);
+    write!(schema_file, "{schema_str}")?;
     Ok(())
 }
