+++ crates/packager/src/package/appimage/mod.rs
@@ -13,11 +13,7 @@ use std::{
 use handlebars::{to_json, Handlebars};
 
 use super::Context;
-use crate::{
-    config::{ConfigExt, ConfigExtInternal},
-    shell::CommandExt,
-    util,
-};
+use crate::{shell::CommandExt, util};
 
 #[tracing::instrument(level = "trace")]
 fn donwload_dependencies(
