+++ crates/packager/src/package/dmg/mod.rs
@@ -8,7 +8,6 @@ use std::{os::unix::fs::PermissionsExt, path::PathBuf, process::Command};
 use super::Context;
 use crate::{
     codesign,
-    config::ConfigExt,
     shell::CommandExt,
     util::{self, download},
 };
